Neon/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
braceexpand==0.1.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
discord.py==2.3.0
frozenlist==1.3.3
idna==3.4
import-expression==1.1.4
jishaku==2.5.1
line-profiler==4.0.3
multidict==6.0.4
pip==22.3.1
prettytable==3.8.0
psutil==5.9.5
pycparser==2.21
PyNaCl==1.5.0
pytz==2023.3
requests==2.31.0
setuptools==65.5.0
six==1.16.0
urllib3==2.0.3
Wavelink==2.5.1
wcwidth==0.2.6
wheel==0.40.0
yarl==1.8.2